Safeguard against non-checkins

+++ omdrivers/lifecycle/iDRAC/RAIDHelper.py
@@ -206,7 +206,7 @@ class RAIDHelper:
         for i in kwargs:
             if i in vdisk.__dict__:
                 if i == 'StripeSize':
-                    kwargs[i] = kwargs[i]/512
+                    kwargs[i] = int(kwargs[i]/512)
                 vdisk.__dict__[i]._value = kwargs[i]
         vdisk._attribs['FQDD'] = vdfqdd
         vdisk.IncludedPhysicalDiskID = ",".join([i.FQDD._value for i in disks])
